Some code to save time
mkgs.py stands for make geometry and source files. It copies the templates attached in models and replaces the half with a number passed in the by the user. mkgs_batch.py does the same thing but for many volume percentages. Pass in a text file as input where each line is a volume percentage, and it will create a geometry and source file for all the volumes. cosima_batch.py takes a directory as an input and applies cosima to each file that ends in a .source extension. repeat.bash might not be useful for all, but it is a bash script that applies cosima multiple times to a specified source file.