Analysis Code
uncertainty.py reads in a directory full of simulation result files and prints out the average, standard deviation in the detection efficiency. fill_vs_efficiency.py reads in a directory of sim files and makes a plot of volume of the tank vs. detection efficency. It also displays the best fit exponential curve.